requirements.txt
fastapi==0.109.0
python-dotenv==1.0.1
pypdf==4.2.0
Pillow==10.2.0
pytest==8.0.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
uvicorn==0.27.0
python-multipart==0.0.6
httpx==0.26.0
ruff==0.1.14
mypy==1.8.0
google-genai==0.2.0
pydantic==2.5.0
vcrpy==6.0.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
boto3==1.34.0
botocore==1.34.0
moto[dynamodb,s3,sqs]==4.2.14
openpyxl==3.1.2
pdf2image==1.17.0
google-api-core==2.15.0